Feature/TAS-38 Add favorite feature (#16)
* initial front changes

* Add favorites in backend

* Adapt frontend

* add id in backend

* Enhances news article data and prediction flow

Adds article ID to data models and API requests to correlate articles for prediction caching.

Includes user authentication token in API requests for secure access.

Retrieves and preserves article like status when fetching news.

Caches prediction results on the backend to minimize redundant ML inferences.

* show prediction if it was predicted before

* Enables users to toggle article favorites

Adds functionality for users to mark articles as favorites.

Introduces new data transfer objects (DTOs) for handling the favorite toggle request and response.

Updates the API interface and news service to include a method for toggling the favorite status of an article.

Modifies the DiscoverFragment to call the new toggleFavorite method in the view model when a user interacts with the like button.

* sync favorites